move decoder to util package
