+++ nexus-app/src/test/java/com/github/nexus/Base64DecoderTest.java
@@ -1,9 +1,12 @@
 package com.github.nexus;
 
 import com.github.nexus.api.exception.DecodingException;
-import java.util.Base64;
-import static org.assertj.core.api.Assertions.assertThat;
+import com.github.nexus.util.Base64Decoder;
 import org.junit.Test;
+
+import java.util.Base64;
+
+import static org.assertj.core.api.Assertions.assertThat;
 
 public class Base64DecoderTest {
